Рефакторит модули pictures и big-picture
