+++ js/picture-gallery/big-picture.js
@@ -1,7 +1,8 @@
-import {isEscape} from '../utils.js';
-import {hideNode, showNode} from '../utils.js';
-import {switchOnModalMode, switchOffModalMode} from '../utils.js';
+import {
+  isEscape, hideNode, showNode, switchOnModalMode, switchOffModalMode
+} from '../utils.js';
 
+const picturesContainerNode = document.querySelector('.pictures');
 const bigPictureNode = document.querySelector('.big-picture');
 const imageNode = bigPictureNode.querySelector('.big-picture__img img');
 const cancelBtnNode = bigPictureNode.querySelector('.big-picture__cancel');
@@ -10,6 +11,8 @@ const socialCaptionNode = bigPictureNode.querySelector('.social__caption');
 const socialCommentsNode = bigPictureNode.querySelector('.social__comments');
 const socialCommentCountNode = bigPictureNode.querySelector('.social__comment-count');
 const commentsLoaderNode = bigPictureNode.querySelector('.comments-loader');
+
+let renderedPictures = [];
 
 hideNode(commentsLoaderNode);
 hideNode(socialCommentCountNode);
@@ -45,20 +48,43 @@ const onDocumentKeydown = (evt) => {
   }
 };
 
+const onPicturesContainerNodeClick = (evt) => {
+  const pictureNode = evt.target.closest('.picture');
+
+  if (pictureNode && evt.currentTarget.contains(pictureNode)) {
+    evt.preventDefault();
+
+    const {id} = pictureNode.dataset;
+    const picture = renderedPictures.find((item) => item.id === Number(id));
+
+    if (picture) {
+      showBigPicture(picture);
+    }
+  }
+};
+
 function hideBigPicture() {
   hideNode(bigPictureNode);
   switchOffModalMode();
 
+  picturesContainerNode.addEventListener('click', onPicturesContainerNodeClick);
   cancelBtnNode.removeEventListener('click', onCancelBtnNodeClick);
   document.removeEventListener('keydown', onDocumentKeydown);
 }
 
-export const showBigPicture = (picture) => {
+export function showBigPicture(picture) {
   showNode(bigPictureNode);
   switchOnModalMode();
 
   updateBigPicture(picture);
 
+  picturesContainerNode.removeEventListener('click', onPicturesContainerNodeClick);
   cancelBtnNode.addEventListener('click', onCancelBtnNodeClick);
   document.addEventListener('keydown', onDocumentKeydown);
+}
+
+export const setRenderedPictures = (pictures) => renderedPictures = pictures;
+
+export const initBigPicture = () => {
+  picturesContainerNode.addEventListener('click', onPicturesContainerNodeClick);
 };
